refactor: extract batch rendering
